Show hide play button into name input into showing moves and cards
