+++ src/components/MovesCounter.jsx
@@ -1,10 +1,30 @@
 import { useState } from 'react';
+import Card from './Card';
+import NameInput from './NameInput'
 
-function MovesCounter() {
+
+function MovesCounter(userNameInput) {
   const [movesCounter, setCount] = useState(0);
+  const updateMovesCounter = () => {
+    setCount = movesCounter + 1;
+  }
 
   return (
-    <div className="counter">Moves: {movesCounter}</div>
+    <>
+      <div className="counter">{movesCounter} Moves: </div>
+      <div className="cards-div">
+          <div className="cards-row-1">
+            <Card />
+            <Card />
+            <Card />
+          </div>
+          <div className="cards-row-2">
+            <Card />
+            <Card />
+            <Card />
+          </div>
+        </div>
+    </>
   )
 }
