+++ src/components/NameInput.jsx
@@ -1,11 +1,30 @@
 import { useState } from 'react';
 import MovesCounter from './MovesCounter';
+import Card from './Card';
 
 function NameInput() {
+  // update name
+  const [userNameInput, setUserNameInput] = useState();
+  const updateName = (nameInput) => {
+    setUserNameInput(nameInput.target.value);
+  };
+
+  // user name input visibility
+  const [userNameInputVisibility, setUserInputVisibilty] = useState(true);
+  const toggleUserInputVisibility = () => {
+    setUserInputVisibilty(false);
+  }
+
+  // const [submitButtonVisibility, setsubmitButtonVisibility] = useState(true)
+
   return (
-    <div class="name-input-div">
-      <label for="nameLabel">What is your name?</label>
-      <input name="nameLabel" placeholder="What is your name?"></input>
+    <div className="name-input-div">
+      <input placeholder="What is your name?" value={userNameInput} onChange={updateName}></input>
+      <p>Hi:  {userNameInput}!</p>
+      <button onClick={toggleUserInputVisibility} >Submit
+      {userNameInputVisibility ? "{}": <MovesCounter />}
+      </button>
+      
     </div>
   )
 }
